depend more on attr_data_structures and move find_attr! there
